+++ app/client/components/svgs/infoIconDark.tsx
@@ -1,10 +1,19 @@
 import React from "react";
 
-export const InfoIconDark = () => (
-  <svg width="17" height="17" viewBox="0 0 17 17" fill="none">
+interface InfoIconDarkProps {
+  fillColor?: string;
+  size?: number;
+}
+export const InfoIconDark = (props: InfoIconDarkProps) => (
+  <svg
+    width={props.size || "17"}
+    height={props.size || "17"}
+    viewBox="0 0 17 17"
+    fill="none"
+  >
     <path
       d="M8.5 0C3.79667 0 0 3.79667 0 8.5C0 13.2033 3.79667 17 8.5 17C13.2033 17 17 13.2033 17 8.5C17 3.79667 13.2033 0 8.5 0ZM9.46333 11.8433V12.2589C9.36889 12.3344 9.23667 12.3911 9.10444 12.4667C8.97222 12.5233 8.82111 12.58 8.67 12.6367C8.51889 12.6933 8.36778 12.7311 8.21667 12.75C8.06556 12.7689 7.91444 12.7878 7.78222 12.7878C7.48 12.7878 7.29111 12.7311 7.17778 12.6178C7.04556 12.5044 6.98889 12.3722 6.98889 12.24C6.98889 12.0889 7.00778 11.9378 7.02667 11.7867C7.04556 11.6356 7.08333 11.4844 7.12111 11.2956L7.91444 7.68778L7.17778 7.51778V7.12111C7.29111 7.08333 7.44222 7.02667 7.65 6.97C7.83889 6.91333 8.04667 6.87556 8.27333 6.83778C8.5 6.8 8.70778 6.76222 8.91556 6.74333C9.12333 6.72444 9.31222 6.70556 9.48222 6.70556L9.69 6.83778L8.63222 11.8433H9.46333ZM9.82222 5.49667C9.65222 5.64778 9.40667 5.72333 9.12333 5.72333C8.85889 5.72333 8.63222 5.64778 8.44333 5.49667C8.27333 5.34556 8.17889 5.15667 8.17889 4.91111C8.17889 4.64667 8.27333 4.43889 8.44333 4.28778C8.61333 4.13667 8.84 4.06111 9.12333 4.06111C9.42556 4.06111 9.65222 4.13667 9.82222 4.28778C9.99222 4.43889 10.0867 4.64667 10.0867 4.91111C10.0678 5.15667 9.99222 5.34556 9.82222 5.49667Z"
-      fill="#333333"
+      fill={props.fillColor || "#333333"}
     />
   </svg>
 );
